requirements.txt
django-telegram-bot==0.1.0
gunicorn==19.1.1
dj-database-url==0.3.0
psycopg2==2.5.4
whitenoise==2.0.6
